Durian Graph Editor Request: Only show handles of selected keyframes
This option for the Graph Editor means that only the handles for selected keyframes get shown in the view, eliminating clutter.

Currently, the selection code isn't aware of this option, so clicking anywhere near where a handle might be may often trigger it to show up. This may/may not be desireable, but we'll see how Lee goes with this first :)