Fix for [#20032] outliner cuts off long names after pasing them
The system that ensures unique ID names chopped off the end
of the name to add a number suffix (.001 or so) regardless of
whether it already had a number or not. In this particular case,
(no suffix number) it now just shaves off the last letter of the
name until it's unique.

I've been testing this patch here for a while, and seems to work
properly, if anyone would like to have a quick look over this
commit and doublecheck, that would be great too.