Bugfix #21041: pressing tab, adds spaces depending on the end of the line
The code used to calculate the number of spaces to insert for a tab (so that indention widths were aligned to multiples of the number of spaces to use) was incorrectly assuming that the line that this was to occur on was blank, using text->curl->len (i.e. the length of the current line).

The code now uses the position of the cursor to determine how many spaces need to be added to it to move it to the next multiple of the tab width.

---

Also, added numpad enter to text-editor keymap for creating new lines for more consistency with user expectations.