Netrender: Fix for API changes (it would be nice if people at least fixed built-in script when breaking things)
Also force post process off.

This is a partial fix for Bug [#21410], there is still the problem that loading back multilayer exr as render results doesn't load the composited result correctly (bug is assigned to Brecht). The results are still ok when downloaded outside of the render engine, so there's a work around for now.