Fix [#21694] text input box last character not editable
Font kerning needs to be set in order to get accurate results out of BLF_width().
Would be nice if this was more automatic, I've added it to a few other places that
seem like they need this though it's a little unclear due to the globals etc. Also
some other minor tweaks when editing text fields.