Bugfix for [#23571] Blender Desktop Icon does not change size under Vista and Windows 7 * Committing new logo, same design as old one, just larger size, by Alexander Kuznetsov . Thank you!
Note: Could only test on Vista. But it should also work on XP machines.
Note 2 (to jesterKing, elubie: Vista uses the winblender.ico also for the blend files, it doesn't use the winblenderfile.ico. But I guess Windows XP use that still?
If not we could remove that, also blenderplayer, is same as blendericon.